pyproject.toml
[project]
name = "tzbot"
version = "0.1.0"
description = "Telegram bot with OpenRouter AI, payments, and persistence"
authors = [{ name = "Your Name" }]
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.13.1",
    "aiohttp>=3.10.10",
    "python-dotenv>=1.0.1",
    "SQLAlchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "uv>=0.4.20",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true
